requirements.txt
asgiref==3.8.1
async-timeout==3.0.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
channels==2.1.2
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cors==1.0.1
cryptography==42.0.8
daphne==2.2.5
distlib==0.3.8
Django==5.0.1
django-auto-logout==0.5.1
django-cors-headers==3.11.0
django-session-timeout==0.1.0
filelock==3.15.4
firebase==4.0.1
future==1.0.0
gcloud==0.18.3
gevent==24.2.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
jwcrypto==1.5.6
oauth2client==4.1.3
pipenv==2024.0.1
platformdirs==4.2.2
proto-plus==1.23.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pyparsing==3.1.2
PyPDF2==3.0.1
Pyrebase4==4.8.0
PySocks==1.7.1
python-jwt==4.1.0
pytz==2024.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.5.0
tldextract==5.1.2
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
virtualenv==20.26.2
zope.event==5.0
zope.interface==6.4.post2
